move project to Poetry
mainly because of lockfile
