+++ packages/model/.gitignore
@@ -1,3 +1,7 @@
-venv/
-out/
+.mypy_cache/
+.ruff_cache/
+.venv/
 data/
+out/
+
+!model/
