--- packages/model/requirements.txt
@@ -1,76 +1,0 @@
-absl-py==2.2.1
-astunparse==1.6.3
-cachetools==5.5.2
-certifi==2025.1.31
-charset-normalizer==3.4.1
-chex==0.1.89
-etils==1.5.2
-flatbuffers==25.2.10
-flax==0.8.5
-fsspec==2025.3.2
-gast==0.6.0
-google-auth==2.38.0
-google-auth-oauthlib==1.2.1
-google-pasta==0.2.0
-grpcio==1.71.0
-h5py==3.13.0
-humanize==4.12.2
-idna==3.10
-importlib_metadata==8.6.1
-importlib_resources==6.5.2
-jax==0.4.30
-jaxlib==0.4.30
-joblib==1.4.2
-keras==2.15.0
-libclang==18.1.1
-Markdown==3.7
-markdown-it-py==3.0.0
-MarkupSafe==3.0.2
-mdurl==0.1.2
-ml-dtypes==0.2.0
-msgpack==1.1.0
-namex==0.0.8
-nest-asyncio==1.6.0
-numpy==1.26.4
-oauthlib==3.2.2
-opt_einsum==3.4.0
-optax==0.2.4
-optree==0.14.1
-orbax-checkpoint==0.4.4
-packaging==23.2
-pandas==2.2.3
-protobuf==4.25.6
-pyasn1==0.6.1
-pyasn1_modules==0.4.2
-Pygments==2.19.1
-python-dateutil==2.9.0.post0
-pytz==2025.2
-PyYAML==6.0.2
-requests==2.32.3
-requests-oauthlib==2.0.0
-rich==14.0.0
-rsa==4.9
-scikit-learn==1.6.1
-scipy==1.13.1
-six==1.17.0
-tensorboard==2.15.2
-tensorboard-data-server==0.7.2
-tensorflow==2.15.0
-tensorflow-decision-forests==1.8.1
-tensorflow-estimator==2.15.0
-tensorflow-hub==0.16.1
-tensorflow-io-gcs-filesystem==0.37.1
-tensorflowjs==4.22.0
-tensorstore==0.1.45
-termcolor==3.0.0
-tf-keras==2.15.0
-threadpoolctl==3.6.0
-toolz==1.0.0
-typing_extensions==4.13.0
-tzdata==2025.2
-urllib3==2.3.0
-Werkzeug==3.1.3
-wrapt==1.14.1
-wurlitzer==3.1.1
-ydf==0.11.0
-zipp==3.21.0
